Added som comments on how to proceed.
The comments explain how I think we should do it. You said you
don't know that much vimscript. I suggest you ':he blabla' for all the
blablas that you don't understand. A critical thing is the difference
between straight highlighting of stuff that matches a regex, and the
definiton of syntax items and the subsequent highlighting of those.
The conceal feature depends on a syntax-defined thing to be highlighted.